Refactor BodyPartMeasurementController::index to use a dedicated Action.
- Extracted logic from `BodyPartMeasurementController::index` to `App\Actions\Measurements\FetchBodyPartMeasurementsIndexAction`.
- The new Action handles fetching, grouping, and formatting of body part measurements.
- Updated `BodyPartMeasurementController` to inject and use the new Action.
- Removed unused private method `getCommonParts` from the controller.
- Fixed `config/database.php` compatibility with PHP 8.3 by replacing `\Pdo\Mysql` constants with `\PDO` constants.
- Formatted code with PSR-12.
